+++ react-rss/src/scss/styles.scss
@@ -11,13 +11,13 @@
 body {
   background-color: #fff;
   font-family: 'EB Garamond', serif;
-  overflow: auto
+  overflow: auto;
 }
 
 a {
   text-decoration: none;
 }
 
-.active{
+.active {
   border-bottom: 2px solid #fff;
 }
